Move ResourcesHost to core/resources module
