+++ src/components/authorization/policy/builder/resource-granter.ts
@@ -1,8 +1,8 @@
 import { mapValues } from 'lodash';
 import { EnhancedResource, many, Many, ResourceShape } from '~/common';
+import type { EnhancedResourceMap } from '~/core';
 import type { ResourceMap } from '../../model/resource-map';
 import { ResourceAction } from '../actions';
-import type { EnhancedResourceMap } from '../resources.host';
 import {
   ChildRelationshipGranter,
   ChildRelationshipsGranter,
