+++ src/core/resources/index.ts
@@ -1,3 +1,4 @@
 export * from './resource-resolver.service';
 export { LoaderFactory } from './loader.registry';
 export * from './resource.loader';
+export * from './resources.host';
