+++ src/core/resources/resource.module.ts
@@ -3,11 +3,22 @@ import { GraphqlModule } from '../graphql';
 import { ResourceLoaderRegistry } from './loader.registry';
 import { ResourceResolver } from './resource-resolver.service';
 import { ResourceLoader } from './resource.loader';
+import { ResourcesHost } from './resources.host';
 
 @Global()
 @Module({
   imports: [GraphqlModule],
-  providers: [ResourceResolver, ResourceLoaderRegistry, ResourceLoader],
-  exports: [ResourceResolver, ResourceLoaderRegistry, ResourceLoader],
+  providers: [
+    ResourcesHost,
+    ResourceResolver,
+    ResourceLoaderRegistry,
+    ResourceLoader,
+  ],
+  exports: [
+    ResourcesHost,
+    ResourceResolver,
+    ResourceLoaderRegistry,
+    ResourceLoader,
+  ],
 })
 export class ResourceModule {}
